+++ web/components/ErrorView.vue
@@ -6,24 +6,22 @@ SPDX-License-Identifier: BSD-3-Clause
 -->
 
 <script setup>
-import {inject, onMounted} from 'vue';
-
 defineProps({
   title: {
     type: String,
     default: () => 'Error',
   },
-  error: {
+  message: {
     type: String,
     default: () => 'An error occurred.',
   },
+  resettable: {
+    type: Boolean,
+    default: () => false,
+  },
 });
-const $cookies = inject('$cookies');
 
-onMounted(() => {
-  $cookies.remove('accessToken');
-  $cookies.remove('exchangeId');
-});
+defineEmits(['reset']);
 </script>
 
 <template>
@@ -33,7 +31,15 @@ onMounted(() => {
     </h2>
 
     <p class="mb-4 text-lg font-normal text-gray-500">
-      {{error}}
+      {{message}}
+    </p>
+
+    <p v-if="resettable">
+      <button
+        class="text-sm font-semibold text-blue-600 hover:text-blue-800"
+        @click="$emit('reset')">
+        {{$t('exchangeReset')}}
+      </button>
     </p>
   </div>
 </template>